[kcalcore] Properly handle exception datetimes for ClockTime events. Contributes to JB#43708
If a recurring event is specified as a ClockTime event, or if an
exception to the event is specified in ClockTime specification,
we need to apply the exception semantics specially.

That is: we should not perform a toLocalZone() conversion before
checking if the exception datetime matches, but just directly
compare the dates and times as-is.